// generated/openapi.ts
export interface paths {
    "/v1/missions/{id}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Query mission
         * @description Returns a single mission specified by its unique ID
         */
        get: operations["getV1MissionsById"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/v1/missions/daily": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Get daily missions
         * @description Returns the list of the daily missions. Limited to 10000. MAY BE SLOW
         */
        get: operations["getV1MissionsDaily"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/v1/missions/filter": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        /**
         * Filter missions
         * @description Returns the list of missions filtered by the specified parameters. All parameters are optional. For data analysis, avoid using this endpoint.
         */
        post: operations["postV1MissionsFilter"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/v1/missions/": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        /**
         * Create a new mission
         * @description Create a new mission. Every field is optional.
         */
        post: operations["postV1Missions"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/v1/missions/sync": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["getV1MissionsSync"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/v1/missions/emit/{wid}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["getV1MissionsEmitByWid"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/v1/missions/emitday/{day}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["getV1MissionsEmitdayByDay"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
}
// ...
export interface operations {
    getV1MissionsById: {
        parameters: {
            query?: never;
            header?: never;
            path: {
                id: number;
            };
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": {
                        id: number;
                        client: {
                            id?: number;
                            name?: string;
                            wayniumid?: string;
                        };
                        type: string | null;
                        status: number;
                        locations: {
                            id?: number;
                            name?: string;
                        }[];
                        passengers: {
                            id?: number;
                            firstname?: string;
                            lastname?: string;
                            name?: string;
                            wayniumid?: string;
                        }[];
                        passengerCount: number;
                        passengerAdditional: {
                            handy?: number;
                            babyseats?: number;
                        };
                        luggages: {
                            small?: number;
                            medium?: number;
                            large?: number;
                        };
                        request_vehicle_category: string;
                        request_vehicle_exact: string;
                        request_language: string;
                        request_currency: string;
                        request_payment: string;
                        note_dispatch: string;
                        note_chauffeur: string;
                        note_passenger: string;
                        date: Record<string, never> | string | number;
                        startTime: string | null;
                        endTime: string | null;
                        partner: {
                            id?: number;
                            name?: string;
                            wayniumid?: string | null;
                        };
                        chauffeur: {
                            id?: number;
                            firstname?: string;
                            lastname?: string;
                            wayniumid?: string;
                        };
                        vehicle: {
                            id?: number;
                            brand?: string;
                            plate?: string;
                            wayniumid?: string;
                        };
                        wayniumid: string;
                        pathCache: string;
                        created_at: Record<string, never>;
                        updated_at: Record<string, never>;
                    } | null;
// ...
                };
            };
        };
    };
// ...
}